requirements_old.txt
flask 
requests 
beautifulsoup4 
openai 
db-sqlite3
docker
python-dotenv
Flask-Mail